Remove default middleware and add html lang
