+++ example/routes/main.tsx
@@ -12,7 +12,11 @@ const navLinks = [
 
 export default () => (
   <>
-    <Helmet defaultTitle="Example" titleTemplate="Example | %s">
+    <Helmet
+      defaultTitle="Example"
+      titleTemplate="Example | %s"
+      htmlAttributes={{ lang: "en" }}
+    >
       <meta charSet="utf-8" />
       <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     </Helmet>
